trival patch

+++ server/webconsole/api/grpc/conn.go
@@ -128,7 +128,7 @@ func AgentBasic(c *gin.Context) {
 		rss = detail["rss"].(float64)
 		for _, v := range as.PluginDetail {
 			// pass the lost ones
-			if time.Now().Unix()-int64(v["last_heartbeat_time"].(float64)) > 180 {
+			if time.Now().Unix()-int64(v["last_heartbeat_time"].(int64)) > 180 {
 				continue
 			}
 			cpu += v["cpu"].(float64)
